[MINOR] Refactor MatrixMetadata/FrameMetadata methods to Metadata
Move common methods from MatrixMetadata and FrameMetadata classes
to abstract parent Metadata class
