+++ src/main/java/org/apache/sysml/api/mlcontext/MatrixMetadata.java
@@ -19,7 +19,6 @@
 
 package org.apache.sysml.api.mlcontext;
 
-import org.apache.sysml.conf.ConfigurationManager;
 import org.apache.sysml.runtime.matrix.MatrixCharacteristics;
 
 /**
@@ -30,11 +29,6 @@ import org.apache.sysml.runtime.matrix.MatrixCharacteristics;
  */
 public class MatrixMetadata extends Metadata {
 
-	private Long numRows = null;
-	private Long numColumns = null;
-	private Long numNonZeros = null;
-	private Integer numRowsPerBlock = null;
-	private Integer numColumnsPerBlock = null;
 	private MatrixFormat matrixFormat;
 
 	public MatrixMetadata() {
@@ -352,152 +346,6 @@ public class MatrixMetadata extends Metadata {
 		this.numNonZeros = matrixCharacteristics.getNonZeros();
 		this.numRowsPerBlock = matrixCharacteristics.getRowsPerBlock();
 		this.numColumnsPerBlock = matrixCharacteristics.getColsPerBlock();
-	}
-
-	/**
-	 * Set the MatrixMetadata fields based on a MatrixCharacteristics object.
-	 *
-	 * @param matrixCharacteristics
-	 *            the matrix metadata as a MatrixCharacteristics object
-	 */
-	public void setMatrixCharacteristics(MatrixCharacteristics matrixCharacteristics) {
-		this.numRows = matrixCharacteristics.getRows();
-		this.numColumns = matrixCharacteristics.getCols();
-		this.numNonZeros = matrixCharacteristics.getNonZeros();
-		this.numRowsPerBlock = matrixCharacteristics.getRowsPerBlock();
-		this.numColumnsPerBlock = matrixCharacteristics.getColsPerBlock();
-	}
-
-	/**
-	 * Obtain the number of rows
-	 *
-	 * @return the number of rows
-	 */
-	public Long getNumRows() {
-		return numRows;
-	}
-
-	/**
-	 * Set the number of rows
-	 *
-	 * @param numRows
-	 *            the number of rows
-	 */
-	public void setNumRows(Long numRows) {
-		this.numRows = numRows;
-	}
-
-	/**
-	 * Obtain the number of columns
-	 *
-	 * @return the number of columns
-	 */
-	public Long getNumColumns() {
-		return numColumns;
-	}
-
-	/**
-	 * Set the number of columns
-	 *
-	 * @param numColumns
-	 *            the number of columns
-	 */
-	public void setNumColumns(Long numColumns) {
-		this.numColumns = numColumns;
-	}
-
-	/**
-	 * Obtain the number of non-zero values
-	 *
-	 * @return the number of non-zero values
-	 */
-	public Long getNumNonZeros() {
-		return numNonZeros;
-	}
-
-	/**
-	 * Set the number of non-zero values
-	 *
-	 * @param numNonZeros
-	 *            the number of non-zero values
-	 */
-	public void setNumNonZeros(Long numNonZeros) {
-		this.numNonZeros = numNonZeros;
-	}
-
-	/**
-	 * Obtain the number of rows per block
-	 *
-	 * @return the number of rows per block
-	 */
-	public Integer getNumRowsPerBlock() {
-		return numRowsPerBlock;
-	}
-
-	/**
-	 * Set the number of rows per block
-	 *
-	 * @param numRowsPerBlock
-	 *            the number of rows per block
-	 */
-	public void setNumRowsPerBlock(Integer numRowsPerBlock) {
-		this.numRowsPerBlock = numRowsPerBlock;
-	}
-
-	/**
-	 * Obtain the number of columns per block
-	 *
-	 * @return the number of columns per block
-	 */
-	public Integer getNumColumnsPerBlock() {
-		return numColumnsPerBlock;
-	}
-
-	/**
-	 * Set the number of columns per block
-	 *
-	 * @param numColumnsPerBlock
-	 *            the number of columns per block
-	 */
-	public void setNumColumnsPerBlock(Integer numColumnsPerBlock) {
-		this.numColumnsPerBlock = numColumnsPerBlock;
-	}
-
-	/**
-	 * Convert the matrix metadata to a MatrixCharacteristics object. If all
-	 * field values are {@code null}, {@code null} is returned.
-	 *
-	 * @return the matrix metadata as a MatrixCharacteristics object, or
-	 *         {@code null} if all field values are null
-	 */
-	public MatrixCharacteristics asMatrixCharacteristics() {
-
-		if ((numRows == null) && (numColumns == null) && (numRowsPerBlock == null) && (numColumnsPerBlock == null)
-				&& (numNonZeros == null)) {
-			return null;
-		}
-
-		long nr = (numRows == null) ? -1 : numRows;
-		long nc = (numColumns == null) ? -1 : numColumns;
-		int nrpb = (numRowsPerBlock == null) ? ConfigurationManager.getBlocksize() : numRowsPerBlock;
-		int ncpb = (numColumnsPerBlock == null) ? ConfigurationManager.getBlocksize() : numColumnsPerBlock;
-		long nnz = (numNonZeros == null) ? -1 : numNonZeros;
-		return new MatrixCharacteristics(nr, nc, nrpb, ncpb, nnz);
-	}
-
-	@Override
-	public String toString() {
-		return "rows: " + fieldDisplay(numRows) + ", columns: " + fieldDisplay(numColumns) + ", non-zeros: "
-				+ fieldDisplay(numNonZeros) + ", rows per block: " + fieldDisplay(numRowsPerBlock)
-				+ ", columns per block: " + fieldDisplay(numColumnsPerBlock);
-	}
-
-	private String fieldDisplay(Object field) {
-		if (field == null) {
-			return "None";
-		} else {
-			return field.toString();
-		}
 	}
 
 	/**
